Delay vote confirmation emails
* Update tests

+++ spec/support/delayed_job_spec_helper.rb
@@ -1,8 +1,10 @@
 module DelayedJobSpecHelper
   def work_off_delayed_jobs
     Delayed::Job.all.each do |job|
-      job.payload_object.perform
-      job.destroy
+     if job.run_at < Time.now
+        job.payload_object.perform
+        job.destroy
+      end
     end
   end
 end
